get ready for publish
