--- clearned/src/lib.rs
@@ -1,7 +1,0 @@
-pub mod prelude {
-    pub use clearned_core::ImmutableIndex;
-    pub use clearned_derive::create_hybrid_index;
-}
-
-#[doc(hidden)]
-pub use clearned_core as private;
