+++ clearned_bench/src/baseline.rs
@@ -1,4 +1,4 @@
-use clearned::private::{
+use limousine_engine::private::{
     search::{lower_bound, upper_bound, OptimalSearch, Search},
     BaseLayer, HybridIndexRangeIterator, ImmutableIndex, Key, NodeLayer, Result, Value,
 };
