--- clearned_core/Cargo.toml
@@ -1,14 +1,0 @@
-[package]
-name = "clearned_core"
-description = "Building blocks for clearned data structures."
-version = "0.1.0"
-edition = "2021"
-
-[dependencies]
-num = "0.4.0"
-trait-set = "0.3.0"
-bytemuck = { version = "1.13", features = ["min_const_generics"] }
-mmap_buffer = "0.1.4"
-
-[features]
-default = []
