--- clearned_derive/Cargo.toml
@@ -1,13 +1,0 @@
-[package]
-name = "clearned_derive"
-description = "Proc macro for building clearned data structures."
-version = "0.1.0"
-edition = "2021"
-
-[lib]
-proc-macro = true
-
-[dependencies]
-syn = { version = "2.0", features = ["full"] }
-quote = "1.0"
-proc-macro2 = "1.0"
