+++ clearned_example/Cargo.toml
@@ -4,4 +4,7 @@ version = "0.1.0"
 edition = "2021"
 
 [dependencies]
-clearned = { path = "../clearned" }
+limousine_engine = { path = "../limousine_engine" }
+
+rand = "0.8.5"
+rand_distr = "0.4.3"
