+++ limousine_engine/Cargo.toml
@@ -1,0 +1,15 @@
+[package]
+name = "limousine_engine"
+version = "0.1.0"
+edition = "2021"
+description = "limousine_engine can reason about a large design continuum of hybrid index designs and materialize an optimal implementation using procedural macros."
+authors = ["Lev Kruglyak <lev.kruglyak2014@gmail.com>"]
+keywords = ["btree", "pgm", "database", "proc_macro", "limousine"]
+license = "Apache-2.0"
+
+# repository = "https://github.com/LevKruglyak/limousine"
+# readme = "crates-io.md"
+
+[dependencies]
+limousine_derive = { path = "../limousine_derive" }
+limousine_core = { path = "../limousine_core" }
